refactor: separates `make-fixes` and `make-comment-fixes`
